requirements.txt
django==5.1.6
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==23.0.0
python-dotenv==1.0.1
# Dependencies for hexgrid and hexdata
numpy==1.26.4
scipy==1.13.0
geojson==3.1.0
h3==3.7.6
shapely==2.0.3 